Add content navigator and rename tree navigator
