+++ src/tui/handlers/mod.rs
@@ -1,2 +1,2 @@
 pub mod keyboard;
-pub mod navigation;
+pub mod navigator;
