+++ src/tui/handlers/navigator/mod.rs
@@ -1,0 +1,2 @@
+pub mod content;
+pub mod tree;
