+++ src/tui/handlers/navigator/tree.rs
